setup redux

+++ .eslintrc.js
@@ -15,6 +15,7 @@ module.exports = {
   plugins: ["react"],
   rules: {
     semi: ["error", "always"],
-    quotes: [2, "single", { "avoidEscape": true }],
+    quotes: [2, "double", { "avoidEscape": true }],
   },
+  
 };
